ritza-real-world-tests/browserless/requirements.txt
playwright==1.48.0
python-dotenv==1.0.0
requests==2.32.3
